New type of plot for the list of daily orders
Added a new plot for the list of daily orders shown by entering 'today' as the input for a ticker. New plot is shown as a table with the rows being marked as the ticker, and the columns being the order type ('Buy', 'Hold', or 'Sell), date of the order, and the current performance of the portfolio following the algorithm (% change between entry portfolio value and current portfolio value).